+++ script/pathinfo.sh
@@ -27,6 +27,7 @@ LOG_FILE="$LOG_PATH/initsvc.log"
 GAMES_PATH="/data/adb/gpu_governor"
 GAMES_FILE="$GAMES_PATH/games.conf"
 GAME_MODE_FILE="$GAMES_PATH/game_mode"
+LOG_LEVEL_FILE="$GAMES_PATH/log_level"
 
 ###############################
 # Log